requirements.txt
flask
openai
pandas
scikit-learn
python-dotenv
flask-cors